requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.0
pygame-ce>=2.4.0
ultralytics>=8.0.0
openai>=1.0.0
numpy>=1.24.0
